refactor ipc locations
