+++ src/storage/src/liquid_array/ipc.rs
@@ -1,30 +1,20 @@
 //! IPC for liquid array.
 
-use std::num::NonZero;
 use std::sync::Arc;
 use std::{any::TypeId, mem::size_of};
 
 use arrow::array::ArrowPrimitiveType;
-use arrow::array::types::UInt16Type;
 use arrow::datatypes::{
     Date32Type, Date64Type, Float32Type, Float64Type, Int8Type, Int16Type, Int32Type, Int64Type,
-    UInt8Type, UInt32Type, UInt64Type,
+    UInt8Type, UInt16Type, UInt32Type, UInt64Type,
 };
 use bytes::Bytes;
 use fsst::Compressor;
 
+use crate::liquid_array::LiquidByteViewArray;
 use crate::liquid_array::LiquidPrimitiveArray;
-use crate::liquid_array::byte_array::ArrowByteType;
-use crate::liquid_array::byte_view_array::{
-    FsstBuffer, LiquidByteViewArray, MemoryBuffer, OffsetView,
-};
-use crate::liquid_array::fix_len_byte_array::ArrowFixedLenByteArrayType;
-use crate::liquid_array::float_array::Exponents;
-use crate::liquid_array::raw::fsst_array::RawFsstBuffer;
-use crate::liquid_array::raw::{BitPackedArray, FsstArray};
-use crate::liquid_array::{IoRange, LiquidPrimitiveType};
+use crate::liquid_array::byte_view_array::MemoryBuffer;
 
-use super::float_array::LiquidFloatType;
 use super::{
     LiquidArrayRef, LiquidByteArray, LiquidDataType, LiquidFixedLenByteArray, LiquidFloatArray,
 };
@@ -48,22 +38,32 @@ const VERSION: u16 = 1;
     +--------------------------------------------------+
 */
 #[repr(C)]
-pub(crate) struct LiquidIPCHeader {
-    magic: [u8; 4],
-    version: u16,
-    logical_type_id: u16,
-    physical_type_id: u16,
-    __padding: [u8; 6],
+pub(super) struct LiquidIPCHeader {
+    pub(super) magic: [u8; 4],
+    pub(super) version: u16,
+    pub(super) logical_type_id: u16,
+    pub(super) physical_type_id: u16,
+    pub(super) __padding: [u8; 6],
 }
 
 const _: () = assert!(size_of::<LiquidIPCHeader>() == LiquidIPCHeader::size());
 
 impl LiquidIPCHeader {
-    pub(crate) const fn size() -> usize {
+    pub(super) const fn size() -> usize {
         16
     }
 
-    fn to_bytes(&self) -> [u8; Self::size()] {
+    pub(super) fn new(logical_type_id: u16, physical_type_id: u16) -> Self {
+        Self {
+            magic: MAGIC.to_le_bytes(),
+            version: VERSION,
+            logical_type_id,
+            physical_type_id,
+            __padding: [0; 6],
+        }
+    }
+
+    pub(super) fn to_bytes(&self) -> [u8; Self::size()] {
         let mut bytes = [0; Self::size()];
         bytes[0..4].copy_from_slice(&self.magic);
         bytes[4..6].copy_from_slice(&self.version.to_le_bytes());
@@ -72,7 +72,7 @@ impl LiquidIPCHeader {
         bytes
     }
 
-    fn from_bytes(bytes: &[u8]) -> Self {
+    pub(super) fn from_bytes(bytes: &[u8]) -> Self {
         if bytes.len() < Self::size() {
             panic!(
                 "value too small for LiquidIPCHeader, expected at least {} bytes, got {}",
@@ -158,864 +158,7 @@ pub fn read_from_bytes(bytes: Bytes, context: &LiquidIPCContext) -> LiquidArrayR
     }
 }
 
-impl<T> LiquidPrimitiveArray<T>
-where
-    T: LiquidPrimitiveType,
-{
-    fn bit_pack_starting_loc() -> usize {
-        let header_size = LiquidIPCHeader::size() + size_of::<T::Native>();
-        (header_size + 7) & !7
-    }
-
-    /*
-    Serialized LiquidPrimitiveArray Memory Layout:
-    +--------------------------------------------------+
-    | LiquidIPCHeader (16 bytes)                       |
-    +--------------------------------------------------+
-
-    +--------------------------------------------------+
-    | reference_value (size_of::<T::Native> bytes)     |  // The reference value (e.g. minimum value)
-    +--------------------------------------------------+
-    | Padding (to 8-byte alignment)                    |  // Padding to ensure 8-byte alignment
-    +--------------------------------------------------+
-
-    +--------------------------------------------------+
-    | BitPackedArray Data                              |
-    +--------------------------------------------------+
-    | [BitPackedArray Header & Bit-Packed Values]      |  // Written by self.bit_packed.to_bytes()
-    +--------------------------------------------------+
-    */
-    pub(crate) fn to_bytes_inner(&self) -> Vec<u8> {
-        // Determine type ID based on the type
-        let physical_type_id = get_physical_type_id::<T>();
-        let logical_type_id = LiquidDataType::Integer as u16;
-        let header = LiquidIPCHeader {
-            magic: MAGIC.to_le_bytes(),
-            version: VERSION,
-            logical_type_id,
-            physical_type_id,
-            __padding: [0; 6],
-        };
-
-        let bit_pack_starting_loc = Self::bit_pack_starting_loc();
-        let mut result = Vec::with_capacity(bit_pack_starting_loc + 256); // Pre-allocate a reasonable size
-
-        // Write header
-        result.extend_from_slice(&header.to_bytes());
-
-        // Write reference value
-        let ref_value_bytes = unsafe {
-            std::slice::from_raw_parts(
-                &self.reference_value as *const T::Native as *const u8,
-                size_of::<T::Native>(),
-            )
-        };
-        result.extend_from_slice(ref_value_bytes);
-        while result.len() < bit_pack_starting_loc {
-            result.push(0);
-        }
-
-        // Let BitPackedArray write the rest of the data
-        self.bit_packed.to_bytes(&mut result);
-
-        result
-    }
-
-    /// Deserialize a LiquidPrimitiveArray from bytes, using zero-copy where possible.
-    ///
-    /// This function creates a LiquidPrimitiveArray from a byte buffer that was
-    /// previously created using `to_bytes()`. It attempts to use zero-copy semantics
-    /// by creating views into the original buffer for the data portions.
-    pub fn from_bytes(bytes: Bytes) -> Self {
-        let header = LiquidIPCHeader::from_bytes(&bytes);
-
-        let physical_id = header.physical_type_id;
-        assert_eq!(physical_id, get_physical_type_id::<T>());
-        let logical_id = header.logical_type_id;
-        assert_eq!(logical_id, LiquidDataType::Integer as u16);
-
-        // Get the reference value
-        let ref_value_ptr = &bytes[LiquidIPCHeader::size()];
-        let reference_value =
-            unsafe { (ref_value_ptr as *const u8 as *const T::Native).read_unaligned() };
-
-        // Skip ahead to the BitPackedArray data
-        let bit_packed_data = bytes.slice(Self::bit_pack_starting_loc()..);
-        let bit_packed = BitPackedArray::<T::UnSignedType>::from_bytes(bit_packed_data);
-
-        Self {
-            bit_packed,
-            reference_value,
-        }
-    }
-}
-
-impl<T> LiquidFloatArray<T>
-where
-    T: LiquidFloatType,
-{
-    /*
-    Serialized LiquidFloatArray Memory Layout:
-    +--------------------------------------------------+
-    | LiquidIPCHeader (16 bytes)                       |
-    +--------------------------------------------------+
-
-    +--------------------------------------------------+
-    | reference_value                                  |
-    | (size_of::<T::SignedIntType::Native> bytes)      |  // The reference value (e.g. minimum value)
-    +--------------------------------------------------+
-    | Padding (to 8-byte alignment)                    |  // Padding to ensure 8-byte alignment
-    +--------------------------------------------------+
-
-    +--------------------------------------------------+
-    | Exponents                                        |
-    +--------------------------------------------------+
-    | e (1 byte)                                       |
-    +--------------------------------------------------+
-    | f (1 byte)                                       |
-    +--------------------------------------------------+
-    | Padding (6 bytes)                                |
-    +--------------------------------------------------+
-
-    +--------------------------------------------------+
-    | Patch Data                                       |
-    +--------------------------------------------------+
-    | patch_length (8 bytes)                           |
-    +--------------------------------------------------+
-    | patch_indices (8 * patch_length btyes)           |
-    +--------------------------------------------------+
-    | patch_values (length *size_of::<T::Native> btyes;|
-    |               8-byte aligned)                    |
-    +--------------------------------------------------+
-
-    +--------------------------------------------------+
-    | BitPackedArray Data                              |
-    +--------------------------------------------------+
-    | [BitPackedArray Header & Bit-Packed Values]      |  // Written by self.bit_packed.to_bytes()
-    +--------------------------------------------------+
-    */
-    pub(crate) fn to_bytes_inner(&self) -> Vec<u8> {
-        // Determine type ID based on the type
-        let physical_type_id = get_physical_type_id::<T>();
-        let logical_type_id = LiquidDataType::Float as u16;
-        let header = LiquidIPCHeader {
-            magic: MAGIC.to_le_bytes(),
-            version: VERSION,
-            logical_type_id,
-            physical_type_id,
-            __padding: [0; 6],
-        };
-
-        let mut result = Vec::with_capacity(256); // Pre-allocate a reasonable size
-
-        // Write header
-        result.extend_from_slice(&header.to_bytes());
-
-        // Write reference value
-        let ref_value_bytes = unsafe {
-            std::slice::from_raw_parts(
-                &self.reference_value as *const <T::SignedIntType as ArrowPrimitiveType>::Native
-                    as *const u8,
-                size_of::<T::Native>(),
-            )
-        };
-        result.extend_from_slice(ref_value_bytes);
-
-        let exponents_starting_loc = (result.len() + 7) & !7;
-        // Insert padding before exponents start
-        while result.len() < exponents_starting_loc {
-            result.push(0);
-        }
-
-        // TODO(): Check if we can do self.exponent as *const u8
-        let exponent_e_bytes =
-            unsafe { std::slice::from_raw_parts(&self.exponent.e as *const u8, 1) };
-        let exponent_f_bytes =
-            unsafe { std::slice::from_raw_parts(&self.exponent.f as *const u8, 1) };
-        // Write exponents and padding
-        result.extend_from_slice(exponent_e_bytes);
-        result.extend_from_slice(exponent_f_bytes);
-        for _i in 0..6 {
-            result.push(0);
-        }
-
-        // Number of bytes occupied by usize is target-dependent; use u64 instead
-        let patch_length = self.patch_indices.len() as u64;
-
-        let patch_length_bytes = unsafe {
-            std::slice::from_raw_parts(&patch_length as *const u64 as *const u8, size_of::<u64>())
-        };
-
-        // Write the patch length
-        result.extend_from_slice(patch_length_bytes);
-
-        if !self.patch_indices.is_empty() {
-            let patch_indices_bytes = unsafe {
-                std::slice::from_raw_parts(
-                    self.patch_indices.as_ptr() as *const u8,
-                    size_of::<u64>() * self.patch_indices.len(),
-                )
-            };
-
-            // Write the patch indices
-            result.extend_from_slice(patch_indices_bytes);
-
-            // Write the patch values
-            let patch_values_bytes = unsafe {
-                std::slice::from_raw_parts(
-                    self.patch_values.as_ptr() as *const u8,
-                    size_of::<T::Native>() * self.patch_indices.len(),
-                )
-            };
-            result.extend_from_slice(patch_values_bytes);
-        }
-        let padding = ((result.len() + 7) & !7) - result.len();
-
-        // Add padding before writing bit-packed array
-        for _i in 0..padding {
-            result.push(0);
-        }
-
-        // Let BitPackedArray write the rest of the data
-        self.bit_packed.to_bytes(&mut result);
-
-        result
-    }
-
-    /// Deserialize a LiquidFloatArray from bytes, using zero-copy where possible.
-    ///
-    /// This function creates a LiquidFloatArray from a byte buffer that was
-    /// previously created using `to_bytes()`. It attempts to use zero-copy semantics
-    /// by creating views into the original buffer for the data portions.
-    pub fn from_bytes(bytes: Bytes) -> Self {
-        let header = LiquidIPCHeader::from_bytes(&bytes);
-
-        let physical_id = header.physical_type_id;
-        assert_eq!(physical_id, get_physical_type_id::<T>());
-        let logical_id = header.logical_type_id;
-        assert_eq!(logical_id, LiquidDataType::Float as u16);
-
-        // Get the reference value
-        let ref_value_ptr = &bytes[LiquidIPCHeader::size()];
-        let reference_value = unsafe {
-            (ref_value_ptr as *const u8 as *const <T::SignedIntType as ArrowPrimitiveType>::Native)
-                .read_unaligned()
-        };
-
-        // Get the exponents
-        let exponents_starting_offset = (LiquidIPCHeader::size()
-            + size_of::<<T::SignedIntType as ArrowPrimitiveType>::Native>()
-            + 7)
-            & !7;
-        let exponents_ptr = &bytes[exponents_starting_offset] as *const u8;
-        let e = unsafe { exponents_ptr.read_unaligned() };
-        let f = unsafe { exponents_ptr.add(1).read_unaligned() };
-        let exp = Exponents { e, f };
-
-        // Get the patch data
-        let patch_data_ptr = unsafe { exponents_ptr.add(8) } as *const u64;
-        let patch_length = unsafe { patch_data_ptr.read_unaligned() } as usize;
-        let mut patch_indices: Vec<u64> = Vec::with_capacity(patch_length);
-        let mut patch_values: Vec<T::Native> = Vec::with_capacity(patch_length);
-        let mut bit_packed_ptr = (patch_data_ptr as *const u8).wrapping_add(size_of::<u64>());
-        if patch_length > 0 {
-            let patch_indices_ptr = bit_packed_ptr as *mut u64;
-            patch_indices.extend_from_slice(unsafe {
-                std::slice::from_raw_parts(patch_indices_ptr, patch_length)
-            });
-            let patch_values_ptr = unsafe { patch_indices_ptr.add(patch_length) } as *mut T::Native;
-            patch_values.extend_from_slice(unsafe {
-                std::slice::from_raw_parts(patch_values_ptr, patch_length)
-            });
-
-            bit_packed_ptr = bit_packed_ptr.wrapping_add(
-                patch_length * size_of::<u64>() + patch_length * size_of::<T::Native>(),
-            );
-        }
-
-        let mut bit_packed_starting_offset =
-            unsafe { bit_packed_ptr.offset_from(&bytes[0] as *const u8) as usize };
-        bit_packed_starting_offset = (bit_packed_starting_offset + 7) & !7;
-
-        // Skip ahead to the BitPackedArray data
-        let bit_packed_data = bytes.slice(bit_packed_starting_offset..);
-        let bit_packed = BitPackedArray::<T::UnsignedIntType>::from_bytes(bit_packed_data);
-
-        Self {
-            exponent: exp,
-            bit_packed,
-            patch_indices,
-            patch_values,
-            reference_value,
-        }
-    }
-}
-
-#[repr(C)]
-struct ByteArrayHeader {
-    key_size: u32,
-    value_size: u32,
-}
-
-impl ByteArrayHeader {
-    const fn size() -> usize {
-        8
-    }
-
-    fn to_bytes(&self) -> [u8; Self::size()] {
-        let mut bytes = [0; Self::size()];
-        bytes[0..4].copy_from_slice(&self.key_size.to_le_bytes());
-        bytes[4..8].copy_from_slice(&self.value_size.to_le_bytes());
-        bytes
-    }
-
-    fn from_bytes(bytes: &[u8]) -> Self {
-        if bytes.len() < Self::size() {
-            panic!(
-                "value too small for ByteArrayHeader, expected at least {} bytes, got {}",
-                Self::size(),
-                bytes.len()
-            );
-        }
-        let key_size = u32::from_le_bytes(bytes[0..4].try_into().unwrap());
-        let value_size = u32::from_le_bytes(bytes[4..8].try_into().unwrap());
-        Self {
-            key_size,
-            value_size,
-        }
-    }
-}
-const _: () = assert!(size_of::<ByteArrayHeader>() == ByteArrayHeader::size());
-
-impl LiquidByteArray {
-    /*
-    Serialized LiquidByteArray Memory Layout:
-
-    +--------------------------------------------------+
-    | LiquidIPCHeader (16 bytes)                       |
-    +--------------------------------------------------+
-    | ByteArrayHeader (8 bytes)                        |  // Contains key_size and value_size
-    +--------------------------------------------------+
-    | BitPackedArray Data (keys)                       |
-    +--------------------------------------------------+
-    | [BitPackedArray Header & Bit-Packed Key Values]  |  // Written by self.keys.to_bytes()
-    +--------------------------------------------------+
-    | Padding (to 8-byte alignment)                    |  // Padding to ensure 8-byte alignment
-    +--------------------------------------------------+
-    | FsstArray Data (values)                          |
-    +--------------------------------------------------+
-    | [FsstArray Data]                                 |  // Written by self.values.to_bytes()
-    +--------------------------------------------------+
-    */
-    pub(crate) fn to_bytes_inner(&self) -> Vec<u8> {
-        // Create a buffer for the final output data, starting with the header
-        let header_size = LiquidIPCHeader::size() + ByteArrayHeader::size();
-        let mut result = Vec::with_capacity(header_size + 1024); // Pre-allocate a reasonable size
-
-        result.resize(header_size, 0);
-
-        // Serialize the BitPackedArray (keys)
-        let keys_start = result.len();
-        self.keys.to_bytes(&mut result);
-        let keys_size = result.len() - keys_start;
-
-        // Add padding to ensure FsstArray starts at an 8-byte aligned position
-        while !result.len().is_multiple_of(8) {
-            result.push(0);
-        }
-
-        // Serialize the FsstArray (values)
-        let values_start = result.len();
-        self.values.to_bytes(&mut result);
-        let values_size = result.len() - values_start;
-
-        // Go back and fill in the header
-        let ipc_header = LiquidIPCHeader {
-            magic: MAGIC.to_le_bytes(),
-            version: VERSION,
-            logical_type_id: LiquidDataType::ByteArray as u16,
-            physical_type_id: self.original_arrow_type as u16,
-            __padding: [0; 6],
-        };
-        let header = &mut result[0..header_size];
-        header[0..LiquidIPCHeader::size()].copy_from_slice(&ipc_header.to_bytes());
-
-        let byte_array_header = ByteArrayHeader {
-            key_size: keys_size as u32,
-            value_size: values_size as u32,
-        };
-        header[LiquidIPCHeader::size()..header_size].copy_from_slice(&byte_array_header.to_bytes());
-
-        result
-    }
-
-    /// Deserialize a LiquidByteArray from bytes, using zero-copy where possible.
-    pub fn from_bytes(bytes: Bytes, compressor: Arc<Compressor>) -> Self {
-        let header_size = LiquidIPCHeader::size() + ByteArrayHeader::size();
-        let header = LiquidIPCHeader::from_bytes(&bytes);
-
-        let byte_array_header =
-            ByteArrayHeader::from_bytes(&bytes[LiquidIPCHeader::size()..header_size]);
-
-        let original_arrow_type = ArrowByteType::from(header.physical_type_id);
-
-        let keys_size = byte_array_header.key_size as usize;
-        let values_size = byte_array_header.value_size as usize;
-
-        // Calculate offsets
-        let keys_start = header_size;
-        let keys_end = keys_start + keys_size;
-
-        if keys_end > bytes.len() {
-            panic!("Keys data extends beyond input buffer");
-        }
-
-        // Ensure values data starts at 8-byte aligned position
-        let values_start = (keys_end + 7) & !7; // Round up to next 8-byte boundary
-        let values_end = values_start + values_size;
-
-        if values_end > bytes.len() {
-            panic!("Values data extends beyond input buffer");
-        }
-
-        // Extract and deserialize components
-        let keys_data = bytes.slice(keys_start..keys_end);
-        let keys = BitPackedArray::<UInt16Type>::from_bytes(keys_data);
-
-        let values_data = bytes.slice(values_start..values_end);
-        let values = FsstArray::from_bytes(values_data, compressor);
-
-        Self {
-            keys,
-            values,
-            original_arrow_type,
-        }
-    }
-}
-
-// Header for LiquidByteViewArray serialization
-#[repr(C)]
-pub(crate) struct ByteViewArrayHeader {
-    keys_size: u32,
-    offset_views_size: u32,
-    shared_prefix_size: u32,
-    fsst_raw_size: u32,
-}
-
-impl ByteViewArrayHeader {
-    pub(crate) const fn size() -> usize {
-        const _: () =
-            assert!(std::mem::size_of::<ByteViewArrayHeader>() == ByteViewArrayHeader::size());
-        16
-    }
-
-    fn to_bytes(&self) -> [u8; Self::size()] {
-        let mut bytes = [0u8; Self::size()];
-        bytes[0..4].copy_from_slice(&self.keys_size.to_le_bytes());
-        bytes[4..8].copy_from_slice(&self.offset_views_size.to_le_bytes());
-        bytes[8..12].copy_from_slice(&self.shared_prefix_size.to_le_bytes());
-        bytes[12..16].copy_from_slice(&self.fsst_raw_size.to_le_bytes());
-        bytes
-    }
-
-    fn from_bytes(bytes: &[u8]) -> Self {
-        if bytes.len() < Self::size() {
-            panic!(
-                "value too small for ByteViewArrayHeader, expected at least {} bytes, got {}",
-                Self::size(),
-                bytes.len()
-            );
-        }
-        let keys_size = u32::from_le_bytes(bytes[0..4].try_into().unwrap());
-        let offset_views_size = u32::from_le_bytes(bytes[4..8].try_into().unwrap());
-        let shared_prefix_size = u32::from_le_bytes(bytes[8..12].try_into().unwrap());
-        let fsst_raw_size = u32::from_le_bytes(bytes[12..16].try_into().unwrap());
-        Self {
-            keys_size,
-            offset_views_size,
-            shared_prefix_size,
-            fsst_raw_size,
-        }
-    }
-}
-
-const _: () = assert!(std::mem::size_of::<ByteViewArrayHeader>() == ByteViewArrayHeader::size());
-
-fn align_up_8(len: usize) -> usize {
-    (len + 7) & !7
-}
-
-impl<B: FsstBuffer> LiquidByteViewArray<B> {
-    /*
-    Serialized LiquidByteViewArray Memory Layout:
-
-    +--------------------------------------------------+
-    | LiquidIPCHeader (16 bytes)                       |
-    +--------------------------------------------------+
-    | ByteViewArrayHeader (16 bytes)                   |  // keys_size, offsets_size, prefix_size, fsst_size
-    +--------------------------------------------------+
-    | Padding (to 8-byte alignment)                    |
-    +--------------------------------------------------+
-    | RawFsstBuffer bytes                              |
-    +--------------------------------------------------+
-    | Padding (to 8-byte alignment)                    |
-    +--------------------------------------------------+
-    | BitPackedArray Data (dictionary_keys)            |
-    +--------------------------------------------------+
-    | [BitPackedArray Header & Values]                 |
-    +--------------------------------------------------+
-    | Padding (to 8-byte alignment)                    |
-    +--------------------------------------------------+
-    | OffsetView bytes (len * 12)                      |
-    +--------------------------------------------------+
-    | Padding (to 8-byte alignment)                    |
-    +--------------------------------------------------+
-    | Shared prefix bytes                              |
-    +--------------------------------------------------+
-    */
-    pub(crate) fn to_bytes_inner(&self) -> Result<Vec<u8>, IoRange> {
-        let header_size = LiquidIPCHeader::size() + ByteViewArrayHeader::size();
-        let mut result = Vec::with_capacity(header_size + 1024);
-        result.resize(header_size, 0);
-
-        // A) Align and serialize RawFsstBuffer first (near the start)
-        while !result.len().is_multiple_of(8) {
-            result.push(0);
-        }
-        let fsst_start = result.len();
-        let fsst_raw_bytes = {
-            let raw = self.fsst_buffer.get_fsst_buffer()?;
-            raw.to_bytes()
-        };
-        result.extend_from_slice(&fsst_raw_bytes);
-        let fsst_raw_size = result.len() - fsst_start;
-
-        // B) Alignment before keys
-        while !result.len().is_multiple_of(8) {
-            result.push(0);
-        }
-
-        // C) Serialize dictionary keys
-        let keys_start = result.len();
-        {
-            let bit_packed = BitPackedArray::<UInt16Type>::from_primitive(
-                self.dictionary_keys.clone(),
-                NonZero::new(16).unwrap(),
-            );
-            bit_packed.to_bytes(&mut result);
-        }
-        let keys_size = result.len() - keys_start;
-
-        // D) Alignment before offset views
-        while !result.len().is_multiple_of(8) {
-            result.push(0);
-        }
-
-        // E) Serialize offset views
-        let offsets_start = result.len();
-        {
-            for ov in self.offset_views.iter() {
-                result.extend_from_slice(&ov.offset().to_le_bytes());
-                result.extend_from_slice(ov.prefix());
-            }
-        }
-        let offset_views_size = result.len() - offsets_start;
-
-        // F) Alignment before shared prefix
-        while !result.len().is_multiple_of(8) {
-            result.push(0);
-        }
-
-        // G) Serialize shared prefix
-        let prefix_start = result.len();
-        result.extend_from_slice(&self.shared_prefix);
-        let shared_prefix_size = result.len() - prefix_start;
-
-        // Prepare headers
-        let ipc = LiquidIPCHeader {
-            magic: MAGIC.to_le_bytes(),
-            version: VERSION,
-            logical_type_id: LiquidDataType::ByteViewArray as u16,
-            physical_type_id: self.original_arrow_type as u16,
-            __padding: [0; 6],
-        };
-        let view_header = ByteViewArrayHeader {
-            keys_size: keys_size as u32,
-            offset_views_size: offset_views_size as u32,
-            shared_prefix_size: shared_prefix_size as u32,
-            fsst_raw_size: fsst_raw_size as u32,
-        };
-
-        // Write headers into reserved space at start
-        let header_slice = &mut result[0..header_size];
-        header_slice[0..LiquidIPCHeader::size()].copy_from_slice(&ipc.to_bytes());
-        header_slice[LiquidIPCHeader::size()..header_size].copy_from_slice(&view_header.to_bytes());
-
-        Ok(result)
-    }
-
-    /// Deserialize a LiquidByteViewArray from bytes.
-    pub fn from_bytes(
-        bytes: Bytes,
-        compressor: Arc<Compressor>,
-    ) -> LiquidByteViewArray<MemoryBuffer> {
-        // 0) Read IPC header and our view header
-        let ipc = LiquidIPCHeader::from_bytes(&bytes);
-        let original_arrow_type = ArrowByteType::from(ipc.physical_type_id);
-        let header_size = LiquidIPCHeader::size() + ByteViewArrayHeader::size();
-        let view_header =
-            ByteViewArrayHeader::from_bytes(&bytes[LiquidIPCHeader::size()..header_size]);
-
-        let mut cursor = header_size;
-
-        // A) Align and read FSST raw buffer first
-        cursor = align_up_8(cursor);
-        let fsst_end = cursor + view_header.fsst_raw_size as usize;
-        if fsst_end > bytes.len() {
-            panic!("FSST raw buffer extends beyond input buffer");
-        }
-        let fsst_raw = bytes.slice(cursor..fsst_end);
-        let raw_buffer = RawFsstBuffer::from_bytes(fsst_raw);
-        cursor = fsst_end;
-
-        // B) Align and read keys
-        cursor = align_up_8(cursor);
-        let keys_end = cursor + view_header.keys_size as usize;
-        if keys_end > bytes.len() {
-            panic!("Keys data extends beyond input buffer");
-        }
-        let keys_data = bytes.slice(cursor..keys_end);
-        let bit_packed = BitPackedArray::<UInt16Type>::from_bytes(keys_data);
-        let dictionary_keys = bit_packed.to_primitive();
-        cursor = keys_end;
-
-        // C) Align and read offset views
-        cursor = align_up_8(cursor);
-        let offsets_end = cursor + view_header.offset_views_size as usize;
-        if offsets_end > bytes.len() {
-            panic!("Offset views data extends beyond input buffer");
-        }
-        let mut offset_views: Vec<OffsetView> = Vec::new();
-        if view_header.offset_views_size > 0 {
-            let chunk = bytes.slice(cursor..offsets_end);
-            if !chunk.len().is_multiple_of(12) {
-                panic!("Offset views bytes not a multiple of 12");
-            }
-            let count = chunk.len() / 12;
-            offset_views.reserve(count);
-            for i in 0..count {
-                let base = i * 12;
-                let off = u32::from_le_bytes(chunk[base..base + 4].try_into().unwrap());
-                let mut prefix: [u8; 8] = [0; 8];
-                prefix.copy_from_slice(&chunk[base + 4..base + 12]);
-                offset_views.push(OffsetView::new(off, prefix));
-            }
-        }
-        cursor = offsets_end;
-
-        // D) Align and read shared prefix
-        cursor = align_up_8(cursor);
-        let prefix_end = cursor + view_header.shared_prefix_size as usize;
-        if prefix_end > bytes.len() {
-            panic!("Shared prefix data extends beyond input buffer");
-        }
-        let shared_prefix = bytes[cursor..prefix_end].to_vec();
-
-        LiquidByteViewArray {
-            dictionary_keys,
-            offset_views: Arc::from(offset_views),
-            fsst_buffer: MemoryBuffer::new(Arc::new(raw_buffer)),
-            original_arrow_type,
-            shared_prefix,
-            compressor,
-        }
-    }
-}
-
-// Header structure for fixed-length byte arrays - contains additional type metadata
-#[repr(C)]
-struct FixedLenByteArrayHeader {
-    key_size: u32,
-    value_size: u32,
-    arrow_type: u8, // 0 for Decimal128, 1 for Decimal256
-    precision: u8,
-    scale: i8,
-    __padding: u8,
-}
-
-impl FixedLenByteArrayHeader {
-    const fn size() -> usize {
-        12 // 4 + 4 + 1 + 1 + 1 + 1
-    }
-
-    fn to_bytes(&self) -> [u8; Self::size()] {
-        let mut bytes = [0; Self::size()];
-        bytes[0..4].copy_from_slice(&self.key_size.to_le_bytes());
-        bytes[4..8].copy_from_slice(&self.value_size.to_le_bytes());
-        bytes[8] = self.arrow_type;
-        bytes[9] = self.precision;
-        bytes[10] = self.scale as u8;
-        bytes
-    }
-
-    fn from_bytes(bytes: &[u8]) -> Self {
-        if bytes.len() < Self::size() {
-            panic!(
-                "value too small for FixedLenByteArrayHeader, expected at least {} bytes, got {}",
-                Self::size(),
-                bytes.len()
-            );
-        }
-        let key_size = u32::from_le_bytes(bytes[0..4].try_into().unwrap());
-        let value_size = u32::from_le_bytes(bytes[4..8].try_into().unwrap());
-        let arrow_type = bytes[8];
-        let precision = bytes[9];
-        let scale = bytes[10] as i8;
-        Self {
-            key_size,
-            value_size,
-            arrow_type,
-            precision,
-            scale,
-            __padding: 0,
-        }
-    }
-}
-
-impl LiquidFixedLenByteArray {
-    /*
-    Serialized LiquidFixedLenByteArray Memory Layout:
-
-    +--------------------------------------------------+
-    | LiquidIPCHeader (16 bytes)                       |
-    +--------------------------------------------------+
-    | FixedLenByteArrayHeader (12 bytes)               |  // Contains key_size, value_size, and type info
-    +--------------------------------------------------+
-    | BitPackedArray Data (keys)                       |
-    +--------------------------------------------------+
-    | [BitPackedArray Header & Bit-Packed Key Values]  |  // Written by self.keys.to_bytes()
-    +--------------------------------------------------+
-    | Padding (to 8-byte alignment)                    |  // Padding to ensure 8-byte alignment
-    +--------------------------------------------------+
-    | FsstArray Data (values)                          |
-    +--------------------------------------------------+
-    | [FsstArray Data]                                 |  // Written by self.values.to_bytes()
-    +--------------------------------------------------+
-    */
-    pub(crate) fn to_bytes_inner(&self) -> Vec<u8> {
-        // Create a buffer for the final output data, starting with the header
-        let header_size = LiquidIPCHeader::size() + FixedLenByteArrayHeader::size();
-        let mut result = Vec::with_capacity(header_size + 1024); // Pre-allocate a reasonable size
-
-        result.resize(header_size, 0);
-
-        // Serialize the BitPackedArray (keys)
-        let keys_start = result.len();
-        self.keys().to_bytes(&mut result);
-        let keys_size = result.len() - keys_start;
-
-        // Add padding to ensure FsstArray starts at an 8-byte aligned position
-        while !result.len().is_multiple_of(8) {
-            result.push(0);
-        }
-
-        // Serialize the FsstArray (values)
-        let values_start = result.len();
-        self.values().to_bytes(&mut result);
-        let values_size = result.len() - values_start;
-
-        // Go back and fill in the header
-        let ipc_header = LiquidIPCHeader {
-            magic: MAGIC.to_le_bytes(),
-            version: VERSION,
-            logical_type_id: LiquidDataType::FixedLenByteArray as u16,
-            physical_type_id: 0, // Not used for FixedLenByteArray, type info is stored in the specialized header
-            __padding: [0; 6],
-        };
-        let header = &mut result[0..header_size];
-        header[0..LiquidIPCHeader::size()].copy_from_slice(&ipc_header.to_bytes());
-
-        // Map the ArrowFixedLenByteArrayType to our header
-        let (arrow_type, precision, scale) = match self.arrow_type() {
-            ArrowFixedLenByteArrayType::Decimal128(p, s) => (0, *p, *s),
-            ArrowFixedLenByteArrayType::Decimal256(p, s) => (1, *p, *s),
-        };
-
-        let fixed_len_byte_array_header = FixedLenByteArrayHeader {
-            key_size: keys_size as u32,
-            value_size: values_size as u32,
-            arrow_type,
-            precision,
-            scale,
-            __padding: 0,
-        };
-        header[LiquidIPCHeader::size()..header_size]
-            .copy_from_slice(&fixed_len_byte_array_header.to_bytes());
-
-        result
-    }
-
-    /// Deserialize a LiquidFixedLenByteArray from bytes, using zero-copy where possible.
-    pub fn from_bytes(bytes: Bytes, compressor: Arc<Compressor>) -> Self {
-        let header_size = LiquidIPCHeader::size() + FixedLenByteArrayHeader::size();
-        let header = LiquidIPCHeader::from_bytes(&bytes);
-
-        // Verify the logical type
-        assert_eq!(
-            header.logical_type_id,
-            LiquidDataType::FixedLenByteArray as u16
-        );
-
-        let fixed_len_header =
-            FixedLenByteArrayHeader::from_bytes(&bytes[LiquidIPCHeader::size()..header_size]);
-
-        // Parse arrow type based on the header
-        let arrow_type = match fixed_len_header.arrow_type {
-            0 => ArrowFixedLenByteArrayType::Decimal128(
-                fixed_len_header.precision,
-                fixed_len_header.scale,
-            ),
-            1 => ArrowFixedLenByteArrayType::Decimal256(
-                fixed_len_header.precision,
-                fixed_len_header.scale,
-            ),
-            _ => panic!(
-                "Unsupported arrow type code: {}",
-                fixed_len_header.arrow_type
-            ),
-        };
-
-        // Calculate offsets
-        let keys_size = fixed_len_header.key_size as usize;
-        let values_size = fixed_len_header.value_size as usize;
-
-        let keys_start = header_size;
-        let keys_end = keys_start + keys_size;
-
-        if keys_end > bytes.len() {
-            panic!("Keys data extends beyond input buffer");
-        }
-
-        // Ensure values data starts at 8-byte aligned position
-        let values_start = (keys_end + 7) & !7; // Round up to next 8-byte boundary
-        let values_end = values_start + values_size;
-
-        if values_end > bytes.len() {
-            panic!("Values data extends beyond input buffer");
-        }
-
-        // Extract and deserialize components
-        let keys_data = bytes.slice(keys_start..keys_end);
-        let keys = BitPackedArray::<UInt16Type>::from_bytes(keys_data);
-
-        let values_data = bytes.slice(values_start..values_end);
-        let values = FsstArray::from_bytes(values_data, compressor);
-
-        Self::from_parts(arrow_type, keys, values)
-    }
-}
-
-fn get_physical_type_id<T: ArrowPrimitiveType>() -> u16 {
+pub(super) fn get_physical_type_id<T: ArrowPrimitiveType>() -> u16 {
     match TypeId::of::<T>() {
         id if id == TypeId::of::<arrow::datatypes::Int8Type>() => 0,
         id if id == TypeId::of::<arrow::datatypes::Int16Type>() => 1,
@@ -1041,6 +184,7 @@ mod tests {
     };
     use arrow_schema::DataType;
 
+    use crate::liquid_array::raw::FsstArray;
     use crate::liquid_array::{LiquidArray, utils::gen_test_decimal_array};
 
     use super::*;
@@ -1267,10 +411,10 @@ mod tests {
 
         assert_eq!(original_arrow.as_ref(), deserialized_arrow.as_ref());
 
-        // Verify the original arrow type is preserved
+        // Verify the original arrow type is preserved via Arrow array types
         assert_eq!(
-            original.original_arrow_type as u8,
-            deserialized.original_arrow_type as u8
+            original.to_arrow_array().data_type(),
+            deserialized.to_arrow_array().data_type()
         );
     }
